refactor: extract startup.py and card_hooks.py, slim __init__.py to 175 lines
- startup.py (~135 lines): backup, migration, asset validation, first
  enemy generation, starter check, item recount
- card_hooks.py (~50 lines): card timer hooks, answer card hooks
- Remove dead globals (new_state, mutator_full_reset, etc.) from __init__
- __init__.py is now a thin orchestrator: imports, singleton registration,
  and one-liner wiring calls to extracted modules

Final result: 1068 lines → 175 lines (84% reduction)
